[DRAFT - D13.20 only] Add Baofeng keyboard mode
This keyboard mode maps the netmon pages 1-6 to keys 1-6 (this should be a lot more intuitive even for historical users of md380tools). Netmon off is 7, netmon clear is 8, syslog dump is 9. Hash is tied to power in a new feature and star is tied to reverse/talkaround mode just like Baofeng H/Ts