[DRAFT - D13.20 only] Add "Set Active" to the contact details
I wanted to put this option at the very top of the menu, however I see that the whole function would have to be reimplemented to avoid getting fucked by the offsets. I may do this in the future depending on my needs.

This should ease the process of creating a codeplug as well, because the contacts list is accessible quickly to swap to a specific talkgroup without having to temporarily remember the ID, create a specific channel or dig through 4 levels of menu. Program your top 1000 and put the most common on top for easy access :)